requirements.txt
asgiref==3.8.1
Django==5.1.5
django-extensions==3.2.3
djangorestframework==3.15.2
pillow==11.1.0
pydot==3.0.4
pyparsing==3.2.1
sqlparse==0.5.3
django-meta>=2.1.0
django-robots>=5.0
pillow>=10.0.0
django-imagekit>=4.1.0
django-seo-js>=0.3.3
